Signal1D/2D tools consolidation
