+++ hyperspy/signal.py
@@ -19,25 +19,12 @@
 import copy
 import os.path
 import warnings
-import math
 import inspect
 from contextlib import contextmanager
 
 import numpy as np
-import numpy.ma as ma
-import scipy.interpolate
-try:
-    from scipy.signal import savgol_filter
-    savgol_imported = True
-except ImportError:
-    savgol_imported = False
 import scipy as sp
 from matplotlib import pyplot as plt
-try:
-    from statsmodels.nonparametric.smoothers_lowess import lowess
-    statsmodels_installed = True
-except:
-    statsmodels_installed = False
 
 from hyperspy.axes import AxesManager
 from hyperspy import io
@@ -50,1259 +37,17 @@ from hyperspy.decorators import auto_replot
 from hyperspy.defaults_parser import preferences
 from hyperspy.misc.io.tools import ensure_directory
 from hyperspy.external.progressbar import progressbar
-from hyperspy.gui.tools import (
-    SpectrumCalibration,
-    SmoothingSavitzkyGolay,
-    SmoothingLowess,
-    SmoothingTV,
-    ButterworthFilter)
-from hyperspy.misc.tv_denoise import _tv_denoise_1d
-from hyperspy.gui.egerton_quantification import BackgroundRemoval
-from hyperspy.decorators import only_interactive
-from hyperspy.decorators import interactive_range_selector
-from scipy.ndimage.filters import gaussian_filter1d
-from hyperspy.misc.spectrum_tools import find_peaks_ohaver
-from hyperspy.misc.image_tools import (shift_image, estimate_image_shift)
-from hyperspy.misc.math_tools import symmetrize, antisymmetrize
 from hyperspy.exceptions import SignalDimensionError, DataDimensionError
 from hyperspy.misc import array_tools
-from hyperspy.misc import spectrum_tools
 from hyperspy.misc import rgb_tools
-from hyperspy.gui.tools import IntegrateArea
-from hyperspy import components
 from hyperspy.misc.utils import underline
 from hyperspy.external.astroML.histtools import histogram
 from hyperspy.drawing.utils import animate_legend
 from hyperspy.misc.hspy_warnings import VisibleDeprecationWarning
 from hyperspy.misc.signal_tools import are_signals_aligned
 
-
-class Signal2DTools(object):
-
-    def estimate_shift2D(self,
-                         reference='current',
-                         correlation_threshold=None,
-                         chunk_size=30,
-                         roi=None,
-                         normalize_corr=False,
-                         sobel=True,
-                         medfilter=True,
-                         hanning=True,
-                         plot=False,
-                         dtype='float',
-                         show_progressbar=None):
-        """Estimate the shifts in a image using phase correlation
-
-        This method can only estimate the shift by comparing
-        bidimensional features that should not change position
-        between frames. To decrease the memory usage, the time of
-        computation and the accuracy of the results it is convenient
-        to select a region of interest by setting the roi keyword.
-
-        Parameters
-        ----------
-
-        reference : {'current', 'cascade' ,'stat'}
-            If 'current' (default) the image at the current
-            coordinates is taken as reference. If 'cascade' each image
-            is aligned with the previous one. If 'stat' the translation
-            of every image with all the rest is estimated and by
-            performing statistical analysis on the result the
-            translation is estimated.
-        correlation_threshold : {None, 'auto', float}
-            This parameter is only relevant when `reference` is 'stat'.
-            If float, the shift estimations with a maximum correlation
-            value lower than the given value are not used to compute
-            the estimated shifts. If 'auto' the threshold is calculated
-            automatically as the minimum maximum correlation value
-            of the automatically selected reference image.
-        chunk_size: {None, int}
-            If int and `reference`=='stat' the number of images used
-            as reference are limited to the given value.
-        roi : tuple of ints or floats (left, right, top bottom)
-             Define the region of interest. If int(float) the position
-             is given axis index(value).
-        sobel : bool
-            apply a sobel filter for edge enhancement
-        medfilter :  bool
-            apply a median filter for noise reduction
-        hanning : bool
-            Apply a 2d hanning filter
-        plot : bool
-            If True plots the images after applying the filters and
-            the phase correlation
-        dtype : str or dtype
-            Typecode or data-type in which the calculations must be
-            performed.
-        show_progressbar : None or bool
-            If True, display a progress bar. If None the default is set in
-            `preferences`.
-
-        Returns
-        -------
-
-        list of applied shifts
-
-        Notes
-        -----
-
-        The statistical analysis approach to the translation estimation
-        when using `reference`='stat' roughly follows [1]_ . If you use
-        it please cite their article.
-
-        References
-        ----------
-
-        .. [1] Schaffer, Bernhard, Werner Grogger, and Gerald
-        Kothleitner. “Automated Spatial Drift Correction for EFTEM
-        Image Series.”
-        Ultramicroscopy 102, no. 1 (December 2004): 27–36.
-
-        """
-        if show_progressbar is None:
-            show_progressbar = preferences.General.show_progressbar
-        self._check_signal_dimension_equals_two()
-        if roi is not None:
-            # Get the indices of the roi
-            yaxis = self.axes_manager.signal_axes[1]
-            xaxis = self.axes_manager.signal_axes[0]
-            roi = tuple([xaxis._get_index(i) for i in roi[2:]] +
-                        [yaxis._get_index(i) for i in roi[:2]])
-
-        ref = None if reference == 'cascade' else \
-            self.__call__().copy()
-        shifts = []
-        nrows = None
-        images_number = self.axes_manager._max_index + 1
-        if reference == 'stat':
-            nrows = images_number if chunk_size is None else \
-                min(images_number, chunk_size)
-            pcarray = ma.zeros((nrows, self.axes_manager._max_index + 1,
-                                ),
-                               dtype=np.dtype([('max_value', np.float),
-                                               ('shift', np.int32,
-                                                (2,))]))
-            nshift, max_value = estimate_image_shift(
-                self(),
-                self(),
-                roi=roi,
-                sobel=sobel,
-                medfilter=medfilter,
-                hanning=hanning,
-                normalize_corr=normalize_corr,
-                plot=plot,
-                dtype=dtype)
-            np.fill_diagonal(pcarray['max_value'], max_value)
-            pbar = progressbar(maxval=nrows * images_number,
-                               disabled=not show_progressbar)
-        else:
-            pbar = progressbar(maxval=images_number,
-                               disabled=not show_progressbar)
-
-        # Main iteration loop. Fills the rows of pcarray when reference
-        # is stat
-        for i1, im in enumerate(self._iterate_signal()):
-            if reference in ['current', 'cascade']:
-                if ref is None:
-                    ref = im.copy()
-                    shift = np.array([0, 0])
-                nshift, max_val = estimate_image_shift(
-                    ref, im, roi=roi, sobel=sobel, medfilter=medfilter,
-                    hanning=hanning, plot=plot,
-                    normalize_corr=normalize_corr, dtype=dtype)
-                if reference == 'cascade':
-                    shift += nshift
-                    ref = im.copy()
-                else:
-                    shift = nshift
-                shifts.append(shift.copy())
-                pbar.update(i1 + 1)
-            elif reference == 'stat':
-                if i1 == nrows:
-                    break
-                # Iterate to fill the columns of pcarray
-                for i2, im2 in enumerate(
-                        self._iterate_signal()):
-                    if i2 > i1:
-                        nshift, max_value = estimate_image_shift(
-                            im,
-                            im2,
-                            roi=roi,
-                            sobel=sobel,
-                            medfilter=medfilter,
-                            hanning=hanning,
-                            normalize_corr=normalize_corr,
-                            plot=plot,
-                            dtype=dtype)
-
-                        pcarray[i1, i2] = max_value, nshift
-                    del im2
-                    pbar.update(i2 + images_number * i1 + 1)
-                del im
-        if reference == 'stat':
-            # Select the reference image as the one that has the
-            # higher max_value in the row
-            sqpcarr = pcarray[:, :nrows]
-            sqpcarr['max_value'][:] = symmetrize(sqpcarr['max_value'])
-            sqpcarr['shift'][:] = antisymmetrize(sqpcarr['shift'])
-            ref_index = np.argmax(pcarray['max_value'].min(1))
-            self.ref_index = ref_index
-            shifts = (pcarray['shift'] +
-                      pcarray['shift'][ref_index, :nrows][:, np.newaxis])
-            if correlation_threshold is not None:
-                if correlation_threshold == 'auto':
-                    correlation_threshold = \
-                        (pcarray['max_value'].min(0)).max()
-                    print("Correlation threshold = %1.2f" %
-                          correlation_threshold)
-                shifts[pcarray['max_value'] <
-                       correlation_threshold] = ma.masked
-                shifts.mask[ref_index, :] = False
-
-            shifts = shifts.mean(0)
-        else:
-            shifts = np.array(shifts)
-            del ref
-        return shifts
-
-    def align2D(self, crop=True, fill_value=np.nan, shifts=None,
-                roi=None,
-                sobel=True,
-                medfilter=True,
-                hanning=True,
-                plot=False,
-                normalize_corr=False,
-                reference='current',
-                dtype='float',
-                correlation_threshold=None,
-                chunk_size=30,
-                interpolation_order=1):
-        """Align the images in place using user provided shifts or by
-        estimating the shifts.
-
-        Please, see `estimate_shift2D` docstring for details
-        on the rest of the parameters not documented in the following
-        section
-
-        Parameters
-        ----------
-        crop : bool
-            If True, the data will be cropped not to include regions
-            with missing data
-        fill_value : int, float, nan
-            The areas with missing data are filled with the given value.
-            Default is nan.
-        shifts : None or list of tuples
-            If None the shifts are estimated using
-            `estimate_shift2D`.
-        interpolation_order: int, default 1.
-            The order of the spline interpolation. Default is 1, linear
-            interpolation.
-
-        Returns
-        -------
-        shifts : np.array
-            The shifts are returned only if `shifts` is None
-
-        Notes
-        -----
-
-        The statistical analysis approach to the translation estimation
-        when using `reference`='stat' roughly follows [1]_ . If you use
-        it please cite their article.
-
-        References
-        ----------
-
-        .. [1] Schaffer, Bernhard, Werner Grogger, and Gerald
-        Kothleitner. “Automated Spatial Drift Correction for EFTEM
-        Image Series.”
-        Ultramicroscopy 102, no. 1 (December 2004): 27–36.
-
-        """
-        self._check_signal_dimension_equals_two()
-        if shifts is None:
-            shifts = self.estimate_shift2D(
-                roi=roi,
-                sobel=sobel,
-                medfilter=medfilter,
-                hanning=hanning,
-                plot=plot,
-                reference=reference,
-                dtype=dtype,
-                correlation_threshold=correlation_threshold,
-                normalize_corr=normalize_corr,
-                chunk_size=chunk_size)
-            return_shifts = True
-        else:
-            return_shifts = False
-        # Translate with sub-pixel precision if necesary
-        for im, shift in zip(self._iterate_signal(),
-                             shifts):
-            if np.any(shift):
-                shift_image(im, -shift,
-                            fill_value=fill_value,
-                            interpolation_order=interpolation_order)
-                del im
-
-        # Crop the image to the valid size
-        if crop is True:
-            shifts = -shifts
-            bottom, top = (int(np.floor(shifts[:, 0].min())) if
-                           shifts[:, 0].min() < 0 else None,
-                           int(np.ceil(shifts[:, 0].max())) if
-                           shifts[:, 0].max() > 0 else 0)
-            right, left = (int(np.floor(shifts[:, 1].min())) if
-                           shifts[:, 1].min() < 0 else None,
-                           int(np.ceil(shifts[:, 1].max())) if
-                           shifts[:, 1].max() > 0 else 0)
-            self.crop_image(top, bottom, left, right)
-            shifts = -shifts
-        if return_shifts:
-            return shifts
-
-    def crop_image(self, top=None, bottom=None,
-                   left=None, right=None):
-        """Crops an image in place.
-
-        top, bottom, left, right : int or float
-
-            If int the values are taken as indices. If float the values are
-            converted to indices.
-
-        See also:
-        ---------
-        crop
-
-        """
-        self._check_signal_dimension_equals_two()
-        self.crop(self.axes_manager.signal_axes[1].index_in_axes_manager,
-                  top,
-                  bottom)
-        self.crop(self.axes_manager.signal_axes[0].index_in_axes_manager,
-                  left,
-                  right)
-
-
-class Signal1DTools(object):
-
-    def shift1D(self,
-                shift_array,
-                interpolation_method='linear',
-                crop=True,
-                fill_value=np.nan,
-                show_progressbar=None):
-        """Shift the data in place over the signal axis by the amount specified
-        by an array.
-
-        Parameters
-        ----------
-        shift_array : numpy array
-            An array containing the shifting amount. It must have
-            `axes_manager._navigation_shape_in_array` shape.
-        interpolation_method : str or int
-            Specifies the kind of interpolation as a string ('linear',
-            'nearest', 'zero', 'slinear', 'quadratic, 'cubic') or as an
-            integer specifying the order of the spline interpolator to
-            use.
-        crop : bool
-            If True automatically crop the signal axis at both ends if
-            needed.
-        fill_value : float
-            If crop is False fill the data outside of the original
-            interval with the given value where needed.
-        show_progressbar : None or bool
-            If True, display a progress bar. If None the default is set in
-            `preferences`.
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        if show_progressbar is None:
-            show_progressbar = preferences.General.show_progressbar
-        self._check_signal_dimension_equals_one()
-        axis = self.axes_manager.signal_axes[0]
-        offset = axis.offset
-        original_axis = axis.axis.copy()
-        pbar = progressbar(
-            maxval=self.axes_manager.navigation_size,
-            disabled=not show_progressbar)
-        for i, (dat, shift) in enumerate(zip(
-                self._iterate_signal(),
-                shift_array.ravel(()))):
-            if np.isnan(shift):
-                continue
-            si = sp.interpolate.interp1d(original_axis,
-                                         dat,
-                                         bounds_error=False,
-                                         fill_value=fill_value,
-                                         kind=interpolation_method)
-            axis.offset = float(offset - shift)
-            dat[:] = si(axis.axis)
-            pbar.update(i + 1)
-
-        axis.offset = offset
-
-        if crop is True:
-            minimum, maximum = np.nanmin(shift_array), np.nanmax(shift_array)
-            if minimum < 0:
-                iminimum = 1 + axis.value2index(
-                    axis.high_value + minimum,
-                    rounding=math.floor)
-                print(iminimum)
-                self.crop(axis.index_in_axes_manager,
-                          None,
-                          iminimum)
-            if maximum > 0:
-                imaximum = axis.value2index(offset + maximum,
-                                            rounding=math.ceil)
-                self.crop(axis.index_in_axes_manager,
-                          imaximum)
-
-    def interpolate_in_between(self, start, end, delta=3,
-                               show_progressbar=None, **kwargs):
-        """Replace the data in a given range by interpolation.
-
-        The operation is performed in place.
-
-        Parameters
-        ----------
-        start, end : {int | float}
-            The limits of the interval. If int they are taken as the
-            axis index. If float they are taken as the axis value.
-
-        delta : {int | float}
-            The windows around the (start, end) to use for interpolation
-
-        show_progressbar : None or bool
-            If True, display a progress bar. If None the default is set in
-            `preferences`.
-
-        All extra keyword arguments are passed to
-        scipy.interpolate.interp1d. See the function documentation
-        for details.
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        if show_progressbar is None:
-            show_progressbar = preferences.General.show_progressbar
-        self._check_signal_dimension_equals_one()
-        axis = self.axes_manager.signal_axes[0]
-        i1 = axis._get_index(start)
-        i2 = axis._get_index(end)
-        if isinstance(delta, float):
-            delta = int(delta / axis.scale)
-        i0 = int(np.clip(i1 - delta, 0, np.inf))
-        i3 = int(np.clip(i2 + delta, 0, axis.size))
-        pbar = progressbar(
-            maxval=self.axes_manager.navigation_size,
-            disabled=not show_progressbar)
-        for i, dat in enumerate(self._iterate_signal()):
-            dat_int = sp.interpolate.interp1d(
-                list(range(i0, i1)) + list(range(i2, i3)),
-                dat[i0:i1].tolist() + dat[i2:i3].tolist(),
-                **kwargs)
-            dat[i1:i2] = dat_int(list(range(i1, i2)))
-            pbar.update(i + 1)
-
-    def _check_navigation_mask(self, mask):
-        if mask is not None:
-            if not isinstance(mask, Signal):
-                raise ValueError("mask must be a Signal instance.")
-            elif mask.axes_manager.signal_dimension not in (0, 1):
-                raise ValueError("mask must be a Signal with signal_dimension "
-                                 "equal to 1")
-            elif (mask.axes_manager.navigation_dimension !=
-                  self.axes_manager.navigation_dimension):
-                raise ValueError("mask must be a Signal with the same "
-                                 "navigation_dimension as the current signal.")
-
-    def estimate_shift1D(self,
-                         start=None,
-                         end=None,
-                         reference_indices=None,
-                         max_shift=None,
-                         interpolate=True,
-                         number_of_interpolation_points=5,
-                         mask=None,
-                         show_progressbar=None):
-        """Estimate the shifts in the current signal axis using
-         cross-correlation.
-
-        This method can only estimate the shift by comparing
-        unidimensional features that should not change the position in
-        the signal axis. To decrease the memory usage, the time of
-        computation and the accuracy of the results it is convenient to
-        select the feature of interest providing sensible values for
-        `start` and `end`. By default interpolation is used to obtain
-        subpixel precision.
-
-        Parameters
-        ----------
-        start, end : {int | float | None}
-            The limits of the interval. If int they are taken as the
-            axis index. If float they are taken as the axis value.
-        reference_indices : tuple of ints or None
-            Defines the coordinates of the spectrum that will be used
-            as eference. If None the spectrum at the current
-            coordinates is used for this purpose.
-        max_shift : int
-            "Saturation limit" for the shift.
-        interpolate : bool
-            If True, interpolation is used to provide sub-pixel
-            accuracy.
-        number_of_interpolation_points : int
-            Number of interpolation points. Warning: making this number
-            too big can saturate the memory
-        mask : Signal of bool data type.
-            It must have signal_dimension = 0 and navigation_shape equal to the
-            current signal. Where mask is True the shift is not computed
-            and set to nan.
-        show_progressbar : None or bool
-            If True, display a progress bar. If None the default is set in
-            `preferences`.
-
-        Returns
-        -------
-        An array with the result of the estimation in the axis units.
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        if show_progressbar is None:
-            show_progressbar = preferences.General.show_progressbar
-        self._check_signal_dimension_equals_one()
-        ip = number_of_interpolation_points + 1
-        axis = self.axes_manager.signal_axes[0]
-        self._check_navigation_mask(mask)
-        if reference_indices is None:
-            reference_indices = self.axes_manager.indices
-
-        i1, i2 = axis._get_index(start), axis._get_index(end)
-        shift_array = np.zeros(self.axes_manager._navigation_shape_in_array,
-                               dtype=float)
-        ref = self.inav[reference_indices].data[i1:i2]
-        if interpolate is True:
-            ref = spectrum_tools.interpolate1D(ip, ref)
-        pbar = progressbar(
-            maxval=self.axes_manager.navigation_size,
-            disabled=not show_progressbar)
-        for i, (dat, indices) in enumerate(zip(
-                self._iterate_signal(),
-                self.axes_manager._array_indices_generator())):
-            if mask is not None and bool(mask.data[indices]) is True:
-                shift_array[indices] = np.nan
-            else:
-                dat = dat[i1:i2]
-                if interpolate is True:
-                    dat = spectrum_tools.interpolate1D(ip, dat)
-                shift_array[indices] = np.argmax(
-                    np.correlate(ref, dat, 'full')) - len(ref) + 1
-            pbar.update(i + 1)
-        pbar.finish()
-
-        if max_shift is not None:
-            if interpolate is True:
-                max_shift *= ip
-            shift_array.clip(-max_shift, max_shift)
-        if interpolate is True:
-            shift_array /= ip
-        shift_array *= axis.scale
-        return shift_array
-
-    def align1D(self,
-                start=None,
-                end=None,
-                reference_indices=None,
-                max_shift=None,
-                interpolate=True,
-                number_of_interpolation_points=5,
-                interpolation_method='linear',
-                crop=True,
-                fill_value=np.nan,
-                also_align=[],
-                mask=None,
-                show_progressbar=None):
-        """Estimate the shifts in the signal axis using
-        cross-correlation and use the estimation to align the data in place.
-
-        This method can only estimate the shift by comparing
-        unidimensional
-        features that should not change the position.
-        To decrease memory usage, time of computation and improve
-        accuracy it is convenient to select the feature of interest
-        setting the `start` and `end` keywords. By default interpolation is
-        used to obtain subpixel precision.
-
-        Parameters
-        ----------
-        start, end : {int | float | None}
-            The limits of the interval. If int they are taken as the
-            axis index. If float they are taken as the axis value.
-        reference_indices : tuple of ints or None
-            Defines the coordinates of the spectrum that will be used
-            as eference. If None the spectrum at the current
-            coordinates is used for this purpose.
-        max_shift : int
-            "Saturation limit" for the shift.
-        interpolate : bool
-            If True, interpolation is used to provide sub-pixel
-            accuracy.
-        number_of_interpolation_points : int
-            Number of interpolation points. Warning: making this number
-            too big can saturate the memory
-        interpolation_method : str or int
-            Specifies the kind of interpolation as a string ('linear',
-            'nearest', 'zero', 'slinear', 'quadratic, 'cubic') or as an
-            integer specifying the order of the spline interpolator to
-            use.
-        crop : bool
-            If True automatically crop the signal axis at both ends if
-            needed.
-        fill_value : float
-            If crop is False fill the data outside of the original
-            interval with the given value where needed.
-        also_align : list of signals
-            A list of Signal instances that has exactly the same
-            dimensions
-            as this one and that will be aligned using the shift map
-            estimated using the this signal.
-        mask : Signal of bool data type.
-            It must have signal_dimension = 0 and navigation_shape equal to the
-            current signal. Where mask is True the shift is not computed
-            and set to nan.
-        show_progressbar : None or bool
-            If True, display a progress bar. If None the default is set in
-            `preferences`.
-
-        Returns
-        -------
-        An array with the result of the estimation. The shift will be
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        See also
-        --------
-        estimate_shift1D
-
-        """
-        if also_align is None:
-            also_align = []
-        self._check_signal_dimension_equals_one()
-        shift_array = self.estimate_shift1D(
-            start=start,
-            end=end,
-            reference_indices=reference_indices,
-            max_shift=max_shift,
-            interpolate=interpolate,
-            number_of_interpolation_points=number_of_interpolation_points,
-            mask=mask,
-            show_progressbar=show_progressbar)
-        for signal in also_align + [self]:
-            signal.shift1D(shift_array=shift_array,
-                           interpolation_method=interpolation_method,
-                           crop=crop,
-                           fill_value=fill_value,
-                           show_progressbar=show_progressbar)
-
-    def integrate_in_range(self, signal_range='interactive'):
-        """ Sums the spectrum over an energy range, giving the integrated
-        area.
-
-        The energy range can either be selected through a GUI or the command
-        line.
-
-        Parameters
-        ----------
-        signal_range : {a tuple of this form (l, r), "interactive"}
-            l and r are the left and right limits of the range. They can be
-            numbers or None, where None indicates the extremes of the interval.
-            If l and r are floats the `signal_range` will be in axis units (for
-            example eV). If l and r are integers the `signal_range` will be in
-            index units. When `signal_range` is "interactive" (default) the
-            range is selected using a GUI.
-
-        Returns
-        -------
-        integrated_spectrum : Signal subclass
-
-        See Also
-        --------
-        integrate_simpson
-
-        Examples
-        --------
-
-        Using the GUI
-
-        >>> s.integrate_in_range()
-
-        Using the CLI
-
-        >>> s_int = s.integrate_in_range(signal_range=(560,None))
-
-        Selecting a range in the axis units, by specifying the
-        signal range with floats.
-
-        >>> s_int = s.integrate_in_range(signal_range=(560.,590.))
-
-        Selecting a range using the index, by specifying the
-        signal range with integers.
-
-        >>> s_int = s.integrate_in_range(signal_range=(100,120))
-
-        """
-
-        if signal_range == 'interactive':
-            self_copy = self.deepcopy()
-            ia = IntegrateArea(self_copy, signal_range)
-            ia.edit_traits()
-            integrated_spectrum = self_copy
-        else:
-            integrated_spectrum = self._integrate_in_range_commandline(
-                signal_range)
-        return integrated_spectrum
-
-    def _integrate_in_range_commandline(self, signal_range):
-        e1 = signal_range[0]
-        e2 = signal_range[1]
-        integrated_spectrum = self.isig[..., e1:e2].integrate1D(-1)
-        return integrated_spectrum
-
-    @only_interactive
-    def calibrate(self):
-        """Calibrate the spectral dimension using a gui.
-
-        It displays a window where the new calibration can be set by:
-        * Setting the offset, units and scale directly
-        * Selection a range by dragging the mouse on the spectrum figure
-         and
-        setting the new values for the given range limits
-
-        Notes
-        -----
-        For this method to work the output_dimension must be 1. Set the
-        view
-        accordingly
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        self._check_signal_dimension_equals_one()
-        calibration = SpectrumCalibration(self)
-        calibration.edit_traits()
-
-    def smooth_savitzky_golay(self,
-                              polynomial_order=None,
-                              window_length=None,
-                              differential_order=0):
-        """Apply a Savitzky-Golay filter to the data in place.
-
-        If `polynomial_order` or `window_length` or `differential_order` are
-        None the method is run in interactive mode.
-
-        Parameters
-        ----------
-        window_length : int
-            The length of the filter window (i.e. the number of coefficients).
-            `window_length` must be a positive odd integer.
-        polynomial_order : int
-            The order of the polynomial used to fit the samples.
-            `polyorder` must be less than `window_length`.
-        differential_order: int, optional
-            The order of the derivative to compute.  This must be a
-            nonnegative integer.  The default is 0, which means to filter
-            the data without differentiating.
-
-        Notes
-        -----
-        More information about the filter in `scipy.signal.savgol_filter`.
-
-        """
-        if not savgol_imported:
-            raise ImportError("scipy >= 0.14 needs to be installed to use"
-                              "this feature.")
-        self._check_signal_dimension_equals_one()
-        if (polynomial_order is not None and
-                window_length is not None):
-            axis = self.axes_manager.signal_axes[0]
-            self.data = savgol_filter(
-                x=self.data,
-                window_length=window_length,
-                polyorder=polynomial_order,
-                deriv=differential_order,
-                delta=axis.scale,
-                axis=axis.index_in_array)
-
-        else:
-            # Interactive mode
-            smoother = SmoothingSavitzkyGolay(self)
-            smoother.differential_order = differential_order
-            if polynomial_order is not None:
-                smoother.polynomial_order = polynomial_order
-            if window_length is not None:
-                smoother.window_length = window_length
-            smoother.edit_traits()
-
-    def smooth_lowess(self,
-                      smoothing_parameter=None,
-                      number_of_iterations=None,
-                      show_progressbar=None):
-        """Lowess data smoothing in place.
-
-        If `smoothing_parameter` or `number_of_iterations` are None the method
-        is run in interactive mode.
-
-        Parameters
-        ----------
-        smoothing_parameter: float or None
-            Between 0 and 1. The fraction of the data used
-            when estimating each y-value.
-        number_of_iterations: int or None
-            The number of residual-based reweightings
-            to perform.
-        show_progressbar : None or bool
-            If True, display a progress bar. If None the default is set in
-            `preferences`.
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-        ImportError if statsmodels is not installed.
-
-        Notes
-        -----
-        This method uses the lowess algorithm from statsmodels. statsmodels
-        is required for this method.
-
-        """
-        if not statsmodels_installed:
-            raise ImportError("statsmodels is not installed. This package is "
-                              "required for this feature.")
-        self._check_signal_dimension_equals_one()
-        if smoothing_parameter is None or number_of_iterations is None:
-            smoother = SmoothingLowess(self)
-            if smoothing_parameter is not None:
-                smoother.smoothing_parameter = smoothing_parameter
-            if number_of_iterations is not None:
-                smoother.number_of_iterations = number_of_iterations
-            smoother.edit_traits()
-        else:
-            self.map(lowess,
-                     exog=self.axes_manager[-1].axis,
-                     frac=smoothing_parameter,
-                     it=number_of_iterations,
-                     is_sorted=True,
-                     return_sorted=False,
-                     show_progressbar=show_progressbar)
-
-    def smooth_tv(self, smoothing_parameter=None, show_progressbar=None):
-        """Total variation data smoothing in place.
-
-        Parameters
-        ----------
-        smoothing_parameter: float or None
-           Denoising weight relative to L2 minimization. If None the method
-           is run in interactive mode.
-        show_progressbar : None or bool
-            If True, display a progress bar. If None the default is set in
-            `preferences`.
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        self._check_signal_dimension_equals_one()
-        if smoothing_parameter is None:
-            smoother = SmoothingTV(self)
-            smoother.edit_traits()
-        else:
-            self.map(_tv_denoise_1d, weight=smoothing_parameter,
-                     show_progressbar=show_progressbar)
-
-    def filter_butterworth(self,
-                           cutoff_frequency_ratio=None,
-                           type='low',
-                           order=2):
-        """Butterworth filter in place.
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        self._check_signal_dimension_equals_one()
-        smoother = ButterworthFilter(self)
-        if cutoff_frequency_ratio is not None:
-            smoother.cutoff_frequency_ratio = cutoff_frequency_ratio
-            smoother.apply()
-        else:
-            smoother.edit_traits()
-
-    def _remove_background_cli(self, signal_range, background_estimator,
-                               show_progressbar=None):
-        from hyperspy.model import Model
-        model = Model(self)
-        model.append(background_estimator)
-        background_estimator.estimate_parameters(
-            self,
-            signal_range[0],
-            signal_range[1],
-            only_current=False)
-        return self - model.as_signal(show_progressbar=show_progressbar)
-
-    def remove_background(
-            self,
-            signal_range='interactive',
-            background_type='PowerLaw',
-            polynomial_order=2,
-            show_progressbar=None):
-        """Remove the background, either in place using a gui or returned as a new
-        spectrum using the command line.
-
-        Parameters
-        ----------
-        signal_range : tuple, optional
-            If this argument is not specified, the signal range has to be
-            selected using a GUI. And the original spectrum will be replaced.
-            If tuple is given, the a spectrum will be returned.
-        background_type : string
-            The type of component which should be used to fit the background.
-            Possible components: PowerLaw, Gaussian, Offset, Polynomial
-            If Polynomial is used, the polynomial order can be specified
-        polynomial_order : int, default 2
-            Specify the polynomial order if a Polynomial background is used.
-        show_progressbar : None or bool
-            If True, display a progress bar. If None the default is set in
-            `preferences`.
-
-        Examples
-        --------
-        >>>> s.remove_background() # Using gui, replaces spectrum s
-        >>>> s2 = s.remove_background(
-                 signal_range=(400,450),
-                 background_type='PowerLaw') #Using cli, returns a spectrum
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        self._check_signal_dimension_equals_one()
-        if signal_range == 'interactive':
-            br = BackgroundRemoval(self)
-            br.edit_traits()
-        else:
-            if background_type == 'PowerLaw':
-                background_estimator = components.PowerLaw()
-            elif background_type == 'Gaussian':
-                background_estimator = components.Gaussian()
-            elif background_type == 'Offset':
-                background_estimator = components.Offset()
-            elif background_type == 'Polynomial':
-                background_estimator = components.Polynomial(polynomial_order)
-            else:
-                raise ValueError(
-                    "Background type: " +
-                    background_type +
-                    " not recognized")
-
-            spectra = self._remove_background_cli(
-                signal_range, background_estimator,
-                show_progressbar=show_progressbar)
-            return spectra
-
-    @interactive_range_selector
-    def crop_spectrum(self, left_value=None, right_value=None,):
-        """Crop in place the spectral dimension.
-
-        Parameters
-        ----------
-        left_value, righ_value: {int | float | None}
-            If int the values are taken as indices. If float they are
-            converted to indices using the spectral axis calibration.
-            If left_value is None crops from the beginning of the axis.
-            If right_value is None crops up to the end of the axis. If
-            both are
-            None the interactive cropping interface is activated
-            enabling
-            cropping the spectrum using a span selector in the signal
-            plot.
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        self._check_signal_dimension_equals_one()
-        self.crop(
-            axis=self.axes_manager.signal_axes[0].index_in_axes_manager,
-            start=left_value, end=right_value)
-
-    @auto_replot
-    def gaussian_filter(self, FWHM):
-        """Applies a Gaussian filter in the spectral dimension in place.
-
-        Parameters
-        ----------
-        FWHM : float
-            The Full Width at Half Maximum of the gaussian in the
-            spectral axis units
-
-        Raises
-        ------
-        ValueError if FWHM is equal or less than zero.
-
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        self._check_signal_dimension_equals_one()
-        if FWHM <= 0:
-            raise ValueError(
-                "FWHM must be greater than zero")
-        axis = self.axes_manager.signal_axes[0]
-        FWHM *= 1 / axis.scale
-        self.data = gaussian_filter1d(
-            self.data,
-            axis=axis.index_in_array,
-            sigma=FWHM / 2.35482)
-
-    @auto_replot
-    def hanning_taper(self, side='both', channels=None, offset=0):
-        """Apply a hanning taper to the data in place.
-
-        Parameters
-        ----------
-        side : {'left', 'right', 'both'}
-        channels : {None, int}
-            The number of channels to taper. If None 5% of the total
-            number of channels are tapered.
-        offset : int
-
-        Returns
-        -------
-        channels
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        # TODO: generalize it
-        self._check_signal_dimension_equals_one()
-        if channels is None:
-            channels = int(round(len(self()) * 0.02))
-            if channels < 20:
-                channels = 20
-        dc = self.data
-        if side == 'left' or side == 'both':
-            dc[..., offset:channels + offset] *= (
-                np.hanning(2 * channels)[:channels])
-            dc[..., :offset] *= 0.
-        if side == 'right' or side == 'both':
-            if offset == 0:
-                rl = None
-            else:
-                rl = -offset
-            dc[..., -channels - offset:rl] *= (
-                np.hanning(2 * channels)[-channels:])
-            if offset != 0:
-                dc[..., -offset:] *= 0.
-        return channels
-
-    def find_peaks1D_ohaver(self, xdim=None, slope_thresh=0, amp_thresh=None,
-                            subchannel=True, medfilt_radius=5, maxpeakn=30000,
-                            peakgroup=10):
-        """Find peaks along a 1D line (peaks in spectrum/spectra).
-
-        Function to locate the positive peaks in a noisy x-y data set.
-
-        Detects peaks by looking for downward zero-crossings in the
-        first derivative that exceed 'slope_thresh'.
-
-        Returns an array containing position, height, and width of each
-        peak.
-
-        'slope_thresh' and 'amp_thresh', control sensitivity: higher
-        values will
-        neglect smaller features.
-
-
-        peakgroup is the number of points around the top peak to search
-        around
-
-        Parameters
-        ---------
-
-
-        slope_thresh : float (optional)
-                       1st derivative threshold to count the peak
-                       default is set to 0.5
-                       higher values will neglect smaller features.
-
-        amp_thresh : float (optional)
-                     intensity threshold above which
-                     default is set to 10% of max(y)
-                     higher values will neglect smaller features.
-
-        medfilt_radius : int (optional)
-                     median filter window to apply to smooth the data
-                     (see scipy.signal.medfilt)
-                     if 0, no filter will be applied.
-                     default is set to 5
-
-        peakgroup : int (optional)
-                    number of points around the "top part" of the peak
-                    default is set to 10
-
-        maxpeakn : int (optional)
-                   number of maximum detectable peaks
-                   default is set to 5000
-
-        subpix : bool (optional)
-                 default is set to True
-
-        Returns
-        -------
-        peaks : structured array of shape _navigation_shape_in_array in which
-        each cell contains an array that contains as many structured arrays as
-        peaks where found at that location and which fields: position, height,
-        width, contains position, height, and width of each peak.
-
-        Raises
-        ------
-        SignalDimensionError if the signal dimension is not 1.
-
-        """
-        # TODO: add scipy.signal.find_peaks_cwt
-        self._check_signal_dimension_equals_one()
-        axis = self.axes_manager.signal_axes[0].axis
-        arr_shape = (self.axes_manager._navigation_shape_in_array
-                     if self.axes_manager.navigation_size > 0
-                     else [1, ])
-        peaks = np.zeros(arr_shape, dtype=object)
-        for y, indices in zip(self._iterate_signal(),
-                              self.axes_manager._array_indices_generator()):
-            peaks[indices] = find_peaks_ohaver(
-                y,
-                axis,
-                slope_thresh=slope_thresh,
-                amp_thresh=amp_thresh,
-                medfilt_radius=medfilt_radius,
-                maxpeakn=maxpeakn,
-                peakgroup=peakgroup,
-                subchannel=subchannel)
-        return peaks
-
-    def estimate_peak_width(self,
-                            factor=0.5,
-                            window=None,
-                            return_interval=False,
-                            show_progressbar=None):
-        """Estimate the width of the highest intensity of peak
-        of the spectra at a given fraction of its maximum.
-
-        It can be used with asymmetric peaks. For accurate results any
-        background must be previously substracted.
-        The estimation is performed by interpolation using cubic splines.
-
-        Parameters
-        ----------
-        factor : 0 < float < 1
-            The default, 0.5, estimates the FWHM.
-        window : None, float
-            The size of the window centred at the peak maximum
-            used to perform the estimation.
-            The window size must be chosen with care: if it is narrower
-            than the width of the peak at some positions or if it is
-            so wide that it includes other more intense peaks this
-            method cannot compute the width and a NaN is stored instead.
-        return_interval: bool
-            If True, returns 2 extra signals with the positions of the
-            desired height fraction at the left and right of the
-            peak.
-        show_progressbar : None or bool
-            If True, display a progress bar. If None the default is set in
-            `preferences`.
-
-        Returns
-        -------
-        width or [width, left, right], depending on the value of
-        `return_interval`.
-
-        """
-        if show_progressbar is None:
-            show_progressbar = preferences.General.show_progressbar
-        self._check_signal_dimension_equals_one()
-        if not 0 < factor < 1:
-            raise ValueError("factor must be between 0 and 1.")
-
-        left, right = (self._get_navigation_signal(),
-                       self._get_navigation_signal())
-        # The signals must be of dtype float to contain np.nan
-        left.change_dtype('float')
-        right.change_dtype('float')
-        axis = self.axes_manager.signal_axes[0]
-        x = axis.axis
-        maxval = self.axes_manager.navigation_size
-        if maxval > 0:
-            pbar = progressbar(maxval=maxval,
-                               disabled=not show_progressbar)
-        for i, spectrum in enumerate(self):
-            if window is not None:
-                vmax = axis.index2value(spectrum.data.argmax())
-                spectrum = spectrum.isig[vmax - window / 2.:vmax + window / 2.]
-                x = spectrum.axes_manager[0].axis
-            spline = scipy.interpolate.UnivariateSpline(
-                x,
-                spectrum.data - factor * spectrum.data.max(),
-                s=0)
-            roots = spline.roots()
-            if len(roots) == 2:
-                left.isig[self.axes_manager.indices] = roots[0]
-                right.isig[self.axes_manager.indices] = roots[1]
-            else:
-                left.isig[self.axes_manager.indices] = np.nan
-                right.isig[self.axes_manager.indices] = np.nan
-            if maxval > 0:
-                pbar.update(i)
-        if maxval > 0:
-            pbar.finish()
-        width = right - left
-        if factor == 0.5:
-            width.metadata.General.title = (
-                self.metadata.General.title + " FWHM")
-            left.metadata.General.title = (
-                self.metadata.General.title + " FWHM left position")
-
-            right.metadata.General.title = (
-                self.metadata.General.title + " FWHM right position")
-        else:
-            width.metadata.General.title = (
-                self.metadata.General.title +
-                " full-width at %.1f maximum" % factor)
-            left.metadata.General.title = (
-                self.metadata.General.title +
-                " full-width at %.1f maximum left position" % factor)
-            right.metadata.General.title = (
-                self.metadata.General.title +
-                " full-width at %.1f maximum right position" % factor)
-        if return_interval is True:
-            return [width, left, right]
-        else:
-            return width
+from hyperspy.misc.signal1D_tools import Signal1DTools
+from hyperspy.misc.signal2D_tools import Signal2DTools
 
 
 class MVATools(object):
@@ -2504,10 +1249,10 @@ class MVATools(object):
         factors.plot(navigator=factors_navigator)
 
 
-class Signal(MVA,
-             MVATools,
-             Signal1DTools,
-             Signal2DTools,):
+class BaseSignal(MVA,
+                 MVATools,
+                 Signal1DTools,
+                 Signal2DTools,):
 
     _record_by = ""
     _signal_type = ""
@@ -4821,6 +3566,14 @@ class Signal(MVA,
         """
         from hyperspy.model import Model
         return Model(self)
+
+
+class Signal(BaseSignal,
+             Signal1DTools,
+             Signal2DTools,):
+    VisibleDeprecationWarning('The Signal class will be deprecated in version\
+                               1.0.0 and replaced with BaseSignal, Signal1D\
+                               and Signal2D')
 
 
 ARITHMETIC_OPERATORS = (
